agent/news_fetcher.py
"""
News article fetching and parsing module
"""

import logging
from typing import List, Dict
from datetime import datetime, timedelta
import requests
from bs4 import BeautifulSoup

logger = logging.getLogger(__name__)


class NewsFetcher:
    """Fetches and parses news articles from various sources"""
    
    def __init__(self, config):
        self.config = config
        self.session = requests.Session()
        self.session.headers.update({
            'User-Agent': 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36'
        })
    
    def fetch_articles(self, num_articles: int = 3) -> List[Dict]:
        """
        Fetch recent news articles
        
        Args:
            num_articles: Number of articles to fetch
            
        Returns:
            List of article dictionaries with keys:
            - title: Original article title
            - content: Article text content
            - url: Source URL
            - source: Source name (e.g., 'CNN')
            - published_date: Publication date
            - category: Suggested category
        """
        logger.info(f"Fetching {num_articles} news articles...")
        
        articles = []
        
        # TODO: Implement actual news fetching
        # Options:
        # 1. Use News API (https://newsapi.org/)
        # 2. RSS feed parsing
        # 3. Direct web scraping with BeautifulSoup
        # 4. Use services like Mercury Parser API
        
        # Placeholder implementation
        articles = self._fetch_from_news_api(num_articles)
        
        return articles[:num_articles]
    
    def _fetch_from_news_api(self, num_articles: int) -> List[Dict]:
        """
        Fetch articles using News API
        
        Note: Requires NEWS_API_KEY in environment
        """
        # TODO: Implement News API integration
        # https://newsapi.org/docs/endpoints/top-headlines
        
        raise NotImplementedError("News API integration not yet implemented")
    
    def _fetch_from_rss(self, feed_url: str) -> List[Dict]:
        """
        Fetch articles from RSS feed
        
        Args:
            feed_url: RSS feed URL
        """
        # TODO: Implement RSS parsing with feedparser
        raise NotImplementedError("RSS parsing not yet implemented")
    
    def _parse_article_content(self, url: str) -> Dict:
        """
        Parse article content from URL
        
        Args:
            url: Article URL
            
        Returns:
            Dictionary with parsed article data
        """
        try:
            response = self.session.get(url, timeout=10)
            response.raise_for_status()
            
            soup = BeautifulSoup(response.content, 'html.parser')
            
            # Extract article content (varies by site)
            # TODO: Implement robust content extraction
            # Consider using newspaper3k or readability-lxml
            
            return {
                'title': soup.find('h1').get_text() if soup.find('h1') else '',
                'content': self._extract_paragraphs(soup),
                'url': url
            }
            
        except Exception as e:
            logger.error(f"Error parsing article {url}: {e}")
            return None
    
    def _extract_paragraphs(self, soup: BeautifulSoup) -> str:
        """Extract main content paragraphs from parsed HTML"""
        paragraphs = soup.find_all('p')
        content = '\n\n'.join([p.get_text().strip() for p in paragraphs if p.get_text().strip()])
        return content
    
    def _categorize_article(self, title: str, content: str) -> str:
        """
        Automatically categorize an article based on content
        
        Args:
            title: Article title
            content: Article content
            
        Returns:
            Category name from config.CATEGORIES
        """
        # TODO: Implement intelligent categorization
        # Could use keyword matching, or ask OpenAI to categorize
        
        # Placeholder: return 'Other'
        return 'Other'
