agent/requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0

# Web scraping and parsing
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Optional: Better article parsing
# newspaper3k>=0.2.8
# readability-lxml>=0.8.1

# Optional: RSS feed parsing
# feedparser>=6.0.10

# Image processing
Pillow>=10.0.0

# Optional: News API client
# newsapi-python>=0.2.7

# Optional: Social media APIs
# tweepy>=4.14.0  # For Twitter/X
# python-instagram>=1.3.2  # For Instagram (deprecated, use Graph API)
